Update Palettize, remove unneeded commented code block
Upated PalettizeInvocation to prioritize Palette Path over Palette Image. If an image is applied to Palette Image, it will be ignored if a path is specified in Palette Path. Empty Palette Path if you wish to use the Palette Image input instead.

Removed RetroizeInvocation node code as it's no longer needed.